backend/pyproject.toml
[tool.poetry]
name = "sugoi-chat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.61.2"
uvicorn = "^0.12.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.16.18"}
bcrypt = "^3.2.0"
jwt = "^1.1.0"
asynctest = "^0.13.0"
httpx = "^0.16.1"
pytest-asyncio = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
